# web/aivigilanza/web/openapi-gpt.yaml
openapi: 3.1.0
info:
  title: FTTH Management System API
  description: |
    API per la gestione dei lavori di installazione fibra ottica (FTTH).
    Permette di creare, visualizzare e gestire lavori, tecnici e squadre.
    
    **Architettura**: Questo endpoint pubblico (servicess.net) fa da proxy 
    verso il backend privato tramite rete Yggdrasil IPv6.
    
    **🔔 Sistema Notifiche Telegram Completo**: Il sistema supporta notifiche automatiche 
    complete via Telegram con webhook e comandi interattivi:
    
    - `POST /works/{id}/notify` - Notifica automatica assegnazione lavoro
    - `PUT /works/{id}/assign/{technician_id}` - Assegna lavoro con notifica
    - `POST /telegram/webhook/{bot_token}` - Webhook per comandi tecnici
    - `POST /telegram/gpt/send` - Invio messaggi da GPT ai tecnici
    
    **Comandi Telegram Disponibili**:
    - `/start` - Registrazione tecnico
    - `/miei_lavori` - Lista lavori assegnati
    - `/accetta <WR>` - Accetta lavoro
    - `/rifiuta <WR>` - Rifiuta lavoro  
    - `/chiudi <WR>` - Chiudi lavoro
    - `/help` - Mostra comandi
    
    **Esempio workflow completo**:
    1. Abilita GPT: `POST /telegram/gpt/status` con `{"enabled": true}`
    2. Crea lavoro: `POST /manual/works` con dati lavoro
    3. Assegna tecnico: `PUT /works/{id}/assign/{technician_id}`
    4. Invia messaggio GPT: `POST /telegram/gpt/send` con `{"target_id": 1, "message": "Nuovo lavoro!"}`
  version: 1.0.0
  contact:
    name: Servicess.net
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /telegram/status:
    get:
      operationId: getTelegramStatus
      summary: Stato bot Telegram
      description: Restituisce informazioni sullo stato del bot Telegram e autorizzazioni GPT
      responses:
        '200':
          description: Stato bot
          content:
            application/json:
              schema:
                type: object
                properties:
                  bot_active:
                    type: boolean
                    description: Se il bot è attivo
                  bot_username:
                    type: string
                    description: Username del bot Telegram
                  webhook_configured:
                    type: boolean
                    description: Se il webhook è configurato
                  gpt_authorized:
                    type: boolean
                    description: Se il GPT è autorizzato a inviare messaggi
                  enable_gpt_button:
                    type: boolean
                    description: Se mostrare il pulsante per abilitare GPT
                  chat_id:
                    type: string
                    description: Chat ID configurato per le notifiche
    post:
      operationId: enableGPTSending
      summary: Abilita invio messaggi da GPT
      description: Abilita l'invio di messaggi Telegram da parte del GPT Actions
      responses:
        '200':
          description: GPT abilitato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "GPT autorizzato a inviare messaggi Telegram"
        '400':
          description: Errore nell'abilitazione
  /telegram/send:
    post:
      operationId: sendTelegramMessage
      summary: Invia messaggio Telegram diretto
      description: Invia un messaggio Telegram usando parametri query (chat_id e text)
      parameters:
        - name: chat_id
          in: query
          required: true
          schema:
            type: string
          description: ID della chat Telegram destinataria
        - name: text
          in: query
          required: true
          schema:
            type: string
          description: Testo del messaggio da inviare
        - name: x-api-key
          in: header
          required: true
          schema:
            type: string
          description: Chiave API per autenticazione
      responses:
        '200':
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message_id:
                    type: integer
                    description: ID del messaggio inviato
                    example: 12345
        '400':
          description: Parametri mancanti
        '500':
          description: Errore nell'invio del messaggio
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...